rearrange demo subpackage
